+++ src/main/java/de/denniskniep/safed/mtls/MtlsAssessment.java
@@ -1,7 +1,6 @@
 package de.denniskniep.safed.mtls;
 
 import de.denniskniep.safed.common.assessment.Assessment;
-import de.denniskniep.safed.common.auth.browser.AuthenticationLog;
 import de.denniskniep.safed.common.auth.browser.Browser;
 import de.denniskniep.safed.common.auth.browser.HttpRequest;
 import de.denniskniep.safed.common.scans.AuthResult;
@@ -14,7 +13,6 @@ import org.htmlunit.util.StringUtils;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Set;
 
 @Service
 public class MtlsAssessment extends Assessment<MtlsScanner, MtlsAppConfig> {
@@ -36,23 +34,27 @@ public class MtlsAssessment extends Assessment<MtlsScanner, MtlsAppConfig> {
     }
 
     @Override
-    protected AuthResult scan(MtlsAppConfig config, MtlsScanner scanner) {
+    protected AuthResult scan(MtlsAppConfig config, MtlsScanner scanner, boolean isBaselineScan) {
         var browserConfig = config.getBrowserConfig();
         try (var browser = Browser.create(browserConfig)){
             var signInRequest = scanner.beforeRequest(new HttpRequest("GET", config.getSignInUrl().toString()));
             Page page = browser.execute(signInRequest);
             return new MtlsAuthResult(config, page.authenticationLog(), page);
         }catch (Exception e) {
-            var sslExceptionMessages = List.of(
-                    "net::ERR_SSL_PROTOCOL_ERROR",
-                    "net::ERR_BAD_SSL_CLIENT_AUTH_CERT"
-            );
 
-            for (var sslExceptionMessage : sslExceptionMessages) {
-                if(StringUtils.containsIgnoreCase(e.getMessage(), sslExceptionMessage)) {
-                    var log = new AuthenticationLog();
-                    var page = new Page(config.getSignInUrl().toString(), "", "", sslExceptionMessage, null, Set.of(), log, null, null);
-                    return new MtlsAuthResult(config, page.authenticationLog(), page);
+            // For a non-baseline scan aborting with SSL Errors is good, so therefore
+            if(!isBaselineScan) {
+                var sslExceptionMessages = List.of(
+                        "net::ERR_SSL_PROTOCOL_ERROR",
+                        "net::ERR_BAD_SSL_CLIENT_AUTH_CERT",
+                        "ERR_SOCKET_NOT_CONNECTED"
+                );
+
+                for (var sslExceptionMessage : sslExceptionMessages) {
+                    if(StringUtils.containsIgnoreCase(e.getMessage(), sslExceptionMessage)) {
+                        var page = Page.from(config.getSignInUrl().toString(), e);
+                        return new MtlsAuthResult(config, page.authenticationLog(), page);
+                    }
                 }
             }