+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed Banner in CLI
 - Compressing Screenshots
 - List all scanners in failed assessment report
+- enhance mTLS scanner reporting for socket level cert errors
 
 ## [1.0.5] - 2026-03-15